원격 DB를 구축하여 학원과 집에서 같은 이미지 사용을 위해 업로드
squash

+++ src/main/java/com/ogani/config/quartz/ImageRemoveTask.java
@@ -58,9 +58,11 @@ public class ImageRemoveTask {
 		
 		File targetDir = Paths.get(uploadPath, getFolderBeforeDay(1), "product").toFile();
 		File[] targetImageList = targetDir.listFiles(file -> oldImagePathList.contains(file.toPath()) == false);
-		for (File targetFile : targetImageList) {
-			log.debug("targetFile = {}", targetFile);
-			targetFile.delete();
+		if (targetImageList.length > 0) {
+			for (File targetFile : targetImageList) {
+				log.debug("targetFile = {}", targetFile);
+				targetFile.delete();
+			}
 		}
 	}
 }
